Validasi Hapus Data Kurban

+++ app/Http/Controllers/KurbanController.php
@@ -108,6 +108,16 @@ class KurbanController extends Controller
      */
     public function destroy(Kurban $kurban)
     {
+        if ($kurban->kurbanHewan->count() >= 1) {
+            flash('Data gagal dihapus, karena data sedang di gunakan di tabel lain')->error();
+            return back();
+        }
+
+        if ($kurban->kurbanPeserta->count() >= 1) {
+            flash('Data gagal dihapus, karena data sedang di gunakan di tabel lain')->error();
+            return back();
+        }
+
         $kurban->delete();
 
         flash('Data berhasil dihapus');
